MB-24152: Allow AccessDeleted doc flag on non-xattr commands
Sync Gateway needs to be able to perform a multi-lookup of a Document
body + XATTR on a deleted document. This however was failing with
INVALID_COMBO.

The problem was that in the move to have both path and doc flags, we
decided that AccessDeleted was a document level flag. Prior to this
AccessDeleted was only allowed to be used on an XATTR operation and
that restriction was carried over. This meant that when checking that
the wholedoc CMD_GET was a valid operation, memcached realised that
the AccessDeleted doc flag was set and the operation wasn’t an XATTR,
which it thought was invalid, hence the response code you got.

To address this we are loosening the restriction upon which operations
AccessDeleted can be used on, thus letting all subdoc operations use
it rather than just ones with the XATTR_PATH flag set as well.

Change-Id: Ie513eaf946c573c81dac57d5ed3f0e6c74e9b9e5
Reviewed-on: http://review.couchbase.org/77374
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>