Don't include buckets not in Ready state in list_buckets
For all "empty" entries in the bucket array we've got a
name of "" which is the "no bucket" everyone have access
to. This caused the response message to have a long
tail of whitespace characters (which would be stripped
on the receiving side as it is the separator between
buckets).

Change-Id: I2ec622aacc73bfd1ea0799a0c5479e27f50aea93
Reviewed-on: http://review.couchbase.org/77421
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>