MB-24148: Increase the backfill buf to avert hangs with large items
Currently we have a bug in backfill buffer accounting wherein
there could be a hang potentially with large items whose size are
almost equal to default max.

This commit addresses the bug by masking it. That is we increase the
default backfill buffer size to a value such that we cannot hit the
bug with default item max size.

We are just masking the bug to reduce the amount of change on this
maintainence branch. On the master we are fully fixing the bug.

Change-Id: I215c496cf92a9e9722291ac9a2ad6d14a0298f5c
Reviewed-on: http://review.couchbase.org/77474
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>