MB-24145: Whitelist error responses
As part of MB-23909 we dropped the user data of an error
response and optionally added in the new error object if
it was defined. It turns out that the ROLLBACK error
contains user data the client needs.

In addtion PROTOCOL_BINARY_RESPONSE_SUBDOC_MULTI_PATH_FAILURE
should be whitelisted.

Change-Id: Ifba79501c10f727ad364ccac20d3fa0c555dd749
Reviewed-on: http://review.couchbase.org/77486
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>