Add chunking to phosphor ioctl dumps
This is a 'first-step' for chunking the tracing dumps via ioctl.

Limitations:

- This still blocks the front-end thread while generating each
chunk. Follow up patches will set up a task on an executor thread
for generating the chunks. This potentially needs work to set up
ioctl to allow for ewouldblock.
- The dumps stick around until manually cleared. Memcached doesn't
currently support scheduling tasks for a 'point-in-time' so something
like that needs adding first before automatic expiry is possible.
- The chunk size isn't configurable. A follow up patch will add this
as a query string style argument.
- The developer experience is somewhat awkward as it currently
requires manually stiching the chunks together. Another follow up
patch will either augment mcctl or add a standalone tool to
automatically handle it.

The workflow for performing a trace is a little clunky as well. It
currently requires:

Start Trace -> Stop Trace -> Start Dump -> Get Chunks -> Clear Dump

An 'on-stop' hook could be registered to phosphor to automatically
start the dump to remove that manual step. This could potentially
synergise quite well with the addition of buffer rotation to
phosphor in future.

Change-Id: I1fb87e29f583c966e57a697cb3dbe3146d8c3cf2
Reviewed-on: http://review.couchbase.org/77497
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>