Generate PCKS#12 file of the client file (needed by macosx)
I've not figured out how to make Transport Security on
MacOSX use a private key read from a single pem file like
we do with OpenSSL, but it may import the key from a
PCKS#12 file.

This patch adds a .p12 version of the same key/cert used
by the test client by using "password" as the pass phrase
for the file.

Change-Id: I609818e119ea9c8dfda77541d9bab7ca2fb8f3f1
Reviewed-on: http://review.couchbase.org/77672
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>