MB-24222: Insert open and close marker in logfiles
This allows for splitting the logfile provided by cbcollect_info
back into the original files.

As an extra bonus the end marker reveals if the log was closed
in a controlled way or not.

A new program is installed mclogsplit which parse a concatinated
logfile and recreates the various segments.

Change-Id: Ie356e6cd16d6702a5072850baa83549cbff792ad
Reviewed-on: http://review.couchbase.org/77769
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>