MB-24055: Change ht_size to same value as ns_server (3079)
In preparation for removing the default value for ht_size from
ns_server where it currently is set (so KV-Engine owns the default),
change the value in configuration.json from 3 to 3079 - matching
ns_server's current default.

Note that this value is expected to be subsequently changed (reduced)
again as part of MB-24055 (to reduce memory overhead).

Change-Id: I1025709029dc2e6807e859c27f6ae6725efb7f53
Reviewed-on: http://review.couchbase.org/77890
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>