CTest: Increase maximum output size recorded to 1MB
While tests run via Jenkins record the output of each test, the size
recorded is extremely limited - only 1KB for test which pass. This can
be problematic when (for example) one is trying to analyse which tests
in a suite were slow - for many of our test programs we run many
(sometimes hundreds) of tests, and hence 1024 characters quickly cuts
off interesting data.

As such, increase the limit to 1MB.

Change-Id: I5b2e5e966507951e01e088829120931a0e3bef11
Reviewed-on: http://review.couchbase.org/77893
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>