MB-24152: Return correct body upon Body+XATTR when no XATTR
When attempting to perform a subdoc multi-lookup to both the (whole)
document body and an XATTR path, the request should partially succeed
even if the XATTR path doesn't exist.

However, there is a bug whereby the body phase of the lookup is
skipped if there were any errors in looking up the XATTR path(s) - the
body lookup is not performed (and an empty result returned).

Correct this - we should attempt the body phase irrespective of the
result of the first XATTR phase.

Change-Id: I673f12a7c42c19fb396d0e6d6b4577d2162eb42c
Reviewed-on: http://review.couchbase.org/77935
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>