MB-24152: Allow CMD_GET in multi-lookup to access non-JSON documents
When using CMD_GET as part of a sub-document multi-lookup request,
allow the document to be read even if it is non-JSON.

The reason for allowing this is twofold:

1) Given CMD_GET returns the whole thing, there's no good reason why
   it needs to be limited to specific datatypes (JSON).

2) If a multi-lookup is performed against a Deleted document which has
   an empty body (i.e. the "traditional" form of a delete), then the
   datatype is 'Raw' for the (non-existent) body. As such, to allow
   multi-lookups against empty, deleted documents we need to skip the
   JSON check for them.
   Note: due to another bug we don't currently support accesing an
   empty deleted body, making this change simplifies the work needed
   to support it (in a follow-up patch).

To allow access to non-JSON (for WholeDoc operations), we defer the
check that the input data is JSON to 'subdoc_operate_one_path' - i.e.
the function which actually does the JSON manipulation - allowing
WholeDoc commands to continue on any type.

Change-Id: Ie5fc27265e721c24ba976ec9b73dd9ed2495a927
Reviewed-on: http://review.couchbase.org/78011
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>