MB-24152: Allow CMD_GET in multi-lookup to access empty, deleted docs
When performing a subdoc multi-lookup with includes a WholeDoc spec
(e.g. CMD_GET + an XATTR lookup), we currently fail if the document
being accessed is deleted and has an empty body - i.e. a "traditional"
delete.

We should allow this to partially succeed (allow access to the XATTR);
to support the use-cases which SyncGW has (they want to issue a
CMD_GET+XATTR, and get back whatever partial information exists for
the key).

To support this, check if the input document for WholeDoc:CMD_GET is
of zero length (indicating there is no body), and if so return
SUBDOC_PATH_ENOENT - indicating that "path" (i.e. the doc body)
doesn't exist.

Change-Id: If35344264c17739b94617b8a47b8a170621a5c90
Reviewed-on: http://review.couchbase.org/78013
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>