MB-23808: Add SUBDOC_MULTI_PATH_FAILURE_DELETED status code
Add a new status code - SUBDOC_MULTI_PATH_FAILURE_DELETED - which is
used to indicate that a multi-lookup operation partially
succeeded/failed, and that the looked up document was deleted.

Without this code, there is an ambiguity when a client makes a request
with the AccessDeleted flag set - was the state of the document alive
or dead?

Change-Id: Ifbdbf4eacf4e429734595314094ca0b1c6a818d2
Reviewed-on: http://review.couchbase.org/78012
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>