MB-22602: Use CMake variables for target paths
Instead of manually specifying the path to various targets, use the
appropriate CMake variable / generator expression. This has the
advantage that if a target is moved to a different location, it will
still be correctly referred to.

In the case of ADD_TEST(), use the variant which specifies the NAME
option - this enables expansion of executable targets to their
absolute path.

Change-Id: I625b9908824969010b7535f2df5a69456d3d5ae4
Reviewed-on: http://review.couchbase.org/78181
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>