MB-22602: Give memcached/include its own CMake project
In preparation for merging the memcached & ep_engine repos, create a
separate CMake project for the memcached/include subdir - the external
API of memcached. This is necessary to avoid a circular dependancy
between the combined (memcached+ep_engine) Project and couchstore:

- couchstore depends on memcached/protocol_binary.h (for couch_dbdump)
- ep_engine depends on couchstore.

By creating a separate CMake project (KVEngineAPI) for the memcached
headers; couchstore can depend on that project, breaking the cycle.

Change-Id: Iff8db50bd4653b985f7b6427084959fff6c2367b
Reviewed-on: http://review.couchbase.org/78189
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Oliver Downard <oliver.downard@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>