MB-24098: Add CMD_DELETE to subdoc multi mutation
This patch allows a CMD_DELETE to be one of the wholedoc body mutations
in a subdoc multi mutation. The goal of this is to allow one to delete a
document and update the system Xattrs in a single atomic operation.

The approach use here is similar to one used by the remove_context which
handles regular CMD_DELETEs. If, after the Xattr mutations are
processed, there are system Xattrs then these are preserved, user Xattrs
are removed and a bucket_store with a document state of Dead is used. If
there are no system Xattrs then we perform a bucket_remove instead.

Change-Id: If70c5f43925a2550f3d6354da30a1987b79d10cc
Reviewed-on: http://review.couchbase.org/78237
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>