MB-24424: Pass correct parameters when making MutationResponse
The isKeyOnly parameter was missing when passed into make_unique. This
patch fixes this issue.  In addition it moves the creation of a
PassiveStream into a separate function.  This enables the function that
creates a PassiveStream to be overridden with one that creates a
MockPassiveStream, which is used in testing.

Change-Id: I58e2e8ca06acba24573d1b4a53aeed85dbcecffa
Reviewed-on: http://review.couchbase.org/78302
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>