Add moxi_hammer, a tool to hammer moxi with traffic
moxi_hammer is a command line utility to drive load to a moxi server
to see how it works under pressure. It works by populating a buffer with
commands it wants to send to the server, and then loops over and over
sending that buffer to the server. It spins up a configurable number
of connections to the server and they all send the same buffer, but
start at a different offset (all by sending a set command).

This is currently a minimum implementation, and there is a fair
amount of @todo's in the code for future enhancements... like
adding more threads ;-)

Change-Id: I06f89651f79a1171e00ea008f86eeb44e2c7e4b3
Reviewed-on: http://review.couchbase.org/78389
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>