MB-24481: Re-initialize terminate_handler once file logger enabled
Previously we initialised our custom terminate handler (for unhandled
exceptions) at the very start of memcached_main, before we have the
file logger available. This has the advantage that we can capture even
very early crashes, but it means details of such crashes are not
written to our log file.

Historically this has lack of information in the log file hasn't made
much material difference, as file logger is generally only setup in
production configurations, and there we have Breakpad which reports
the same information. However, now record details of the uncaught
exception in terminate_handler, and hence that reports useful
informaiton not available to breakpad.

As such, add the ability to re-install the terminate handler with a
different logger. Re-install the terminate handler immediatly after
the file logger is setup.

Change-Id: I6bc9d7b4d3d78cb829af426703e6ba2aedb9976b
Reviewed-on: http://review.couchbase.org/78402
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>