MB-23377: Raise GetValue out of callback
This commit changes the kvstore get and getheader interfaces to pass the
value directly via a GetValue rather than a callback which saves a
GetValue. This can be done because all the work is done synchronously.
It also simplifies the interface to allow for additional changes to
GetValue in later patches.

Change-Id: I1c8ccc4e8af39d533ae2a92a1a777d6899b717ea
Reviewed-on: http://review.couchbase.org/78600
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>