MB-24582: emfile test: Handle EOF correctly when searching log file
logger_emfile_test doesn't correctly handle when
wait_for_log_to_contain() encounters end-of-file before the searched
for string is printed - it can hang.

The problem is when reading from an FD which another actor is writing
to (i.e. the log file), if EOF is encountered during reading then any
subsequent reads will also return EOF, even if more data has been
written. It is necessary to clear the EOF flag, which can be done
using fseek() to 'seek' to the same location we are currently
at. Subsequent readas can then see the new data.

Change-Id: I50423409886481dfcac3641c2663f2dc4471e5d0
Reviewed-on: http://review.couchbase.org/78781
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>