MB-24607: Disable autoselect of default bucket
The default bucket should only be selected if the bucket
user is set up without a password.

From MB-23456:

On fresh installed clusters the plan is to not allow anonymous
access to the default bucket. In upgraded clusters, as part of
the upgrade, we will generate a user called "default" with an
empty password so applications against existing clusters will
continue to work post upgrade. We may not have done a great job
letting folks know that this is going to happen, so I thought
I better call it out explicitly in this ticket. At any rate,
this isn't a bug. I'll resolve this ticket, but please feel
free to reply back if you'd like to talk about it -
particularly if you think it's a bad idea.

Change-Id: I47741ea99091c27d59a29455dcab9c8cc8fcd0c3
Reviewed-on: http://review.couchbase.org/78815
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>