MB-24470: [Refactor] Implement insertFromWarmup() in EPVBucket only
Insert from warmup is done only in the case of persistent vbuckets.
Hence it is appropriate to implement it in EPVBucket only.

For now, warmup module contains a reference to base KVBucket
and hence a reference to the base class VBucket. Hence we have to
use dynamic_cast to call the EPVBucket function.

Change-Id: I0a53bc3c3385f0b3e8359033f02b1a9c6a1a922e
Reviewed-on: http://review.couchbase.org/78857
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>