MB-24470: Do not add temp items to the sequence list
In certain cases of delete_with_meta() we add a short lived temp item
in the hashtable and then immediately generate a valid seqno to it,
hence making it a non-temp item. In Ephemeral buckets, this item
while temp is put onto the sequence list with a negative seqno (temp
items have negative seqno in ep-engine). Having a negative seqno
item in the sequence list can cause problems for range (sequential)
readers like DCP, tombstone purge.

This commit ensures that we do not add to temp item on the
sequence list. Temp item is still added on the hash table of legacy
reasons. When the temp item is updated, that is made "non-temp", we
put it onto the sequence list under the "list write lock" and
generate a valid sequence number to it.

The commit also adds some module tests for the same.

Change-Id: I8a98cee401af11f59858d0a2f39cff3cbd540cbf
Reviewed-on: http://review.couchbase.org/78689
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>