Unset env variable in testsuite teardown
The no-default-user test suite exports an environment variable
with completely disable the auto association of the default
bucket. If you run make test it spins up memcached_testapp
with only a few test suites at a time and terminates when
it is done, and the no-default-user testsuite runs by its
own. Without unsetting the variable in the tear down of the
testsuite all tests running after the no-default-user
testsuite will fail if they expect auto association with
the default bucket.

Change-Id: I80ce2e6c2df564db2ef27e00a0f123cf417844ad
Reviewed-on: http://review.couchbase.org/78950
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>