MB-24654: Don't crash if no error map provided
The loading of error maps is done very early, before we have the
logger initialized. If it fails it throws an exception which is
unhandled, and we have little means to log it (no Breakpad, no log
file).

This situation occurs if no valid path to the error maps is specified
- for example when running memcached.exe from the command line with no
config file. This isn't a supported configuration, but it /is/ useful
when debugging environmental issues to be able to run memcached
standalone to check it at least starts (and the version it is).

To assist with this, (1) defer the actual loading of the error map
until we have the logger enabled, and (2) add a try/catch around
loadErrorMap, which can print an explicit error (and exit with a
non-zero status).

With this change if you run memcached standalone it'll fail later
(actually before error map loaded), but with an explicit message:

    $ install/bin/memcached
    Couchbase version 5.0.0-9999 starting.
    RBAC file not specified

Change-Id: Ib6e4e36bfd393aeaa56e8c055804a078506c436f
Reviewed-on: http://review.couchbase.org/78964
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>