MB-23116: get_if() should do a meta_only fetch on first pass
get_if() was added to improve the efficiency of supporting documents
with XATTRs - when a SET request is received it is necessary to check
if the document has any existing XATTRs associated with it, and if so
fetch them so they can be merged into the new value.

To minimise the performance impact of this (particulary in
full-eviction mode), we don't want to have to fetch the whole document
from disk, if it turns out it doesn't have any XATTRs. This is
expected to be the common case (at least initially), as XATTRs are new
and only expected to be used by SyncGW.

Therefore adjust the logic in get_if() for Full-eviction to perform a
meta-only bgfetch on the first pass- i.e. if the document's metadata
isn't resident, then only retrieve the metadata and not the body
also. The specified filter predicate is then applied to that meta, and
if true then perform a full bgfetch.

This also addresses an accounting / stats issue (see linked MB) - as
get_if always performed a full bg_fetch, the cache miss ratio
(bg_fetches / cmd_get) was inflated; as essentially every cmd_set to a
non-resident item was *also* doing a bg fetch. To be more accurate in
a full-eviction scenario we'd have to change the formula to
(bg_fetches / cmd_get + cmd_set), but that would have been less
accurate in value-eviction.

Instead, by changing the bg_fetch to a meta one (which is counted as
meta_bg_fetched) this 'inflaction' is removed; but we still accurately
count the number of full bg_fetches required (say if get_if's filter
matched, and the body needed to be fetched).

Local testing (MacBookPro, 1k binary items, Full eviction, 25%
resident) shows a ~12% speedup in pillowfight with this change:

    cbc-pillowfight -U couchbase://127.0.0.1:12000/default -I 750000 -u default -P asdasd --no-population  --docs random1 --docs random2

Note: This 'optimization' is a performance tradeoff based on the
      expected ratio of filter hit/misses under Full Eviction:
      Currently get_if filter is used only for checking if a Document
      has XATTRs, and so the assumption is such documents will be rare
      in general, hence we first perform a meta bg fetch (instead of
      full meta+value) as the value is not expected to be needed.

      If however this assumption fails, and actually it is more common
      to have documents which match the filter, then this tradeoff
      should be re-visited, as we'll then need to go to disk a
      *second* time for the value (whereas we could have just fetched
      the meta+value in the first place).

Change-Id: I4df2c2747b3af857ca71af5f427e2d0f738c0ea3
Reviewed-on: http://review.couchbase.org/78975
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>