Add ability to mock ProcessClock in ExecutorPool
This makes use of platform's new ProcessClockSource interface to
make it possible to inject a different clock source into the
ExecutorPool which allows for changing the time during tests
in a reliable, efficient manner.

This patch also exposes queue sizes for the various executor
queues. This is intended for use in the tests to verify which
queue a task resides in but is also a useful stat to be exposed
in future.

Finally this patch updates the FutureExecuton test in the
executor test suite to make use of a mock clock source to allow
for a more robust and thorough test of the futureq behaviour.

Change-Id: Ie2c164ac34fe7d66b8210b5c5188837788b775d6
Reviewed-on: http://review.couchbase.org/79022
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>