Refactor: Reduce use of output buffer
All of the permutations of the get commands used to add the
flags to the output buffer, but they could be sent from the
item directly as the reference to the item is kept for the
duration of the commands send phase.

For append/prepend the extras field is only 16 bytes and
could just be kept in the context object instead (and have
the same lifetime as the command), instead of using a buffer
with an "unknown" lifetime.

Change-Id: I6c535d394b320be9ef638a1bb99baeaaf56eb18d
Reviewed-on: http://review.couchbase.org/79091
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>