Refactor: Simplify mcbp_write_and_free logic
Previously we set the connections write.curr and size
to point to a different buffer so that we could add it
to the iovector in a different task. It is easier to
add them to the iovector initially, and then jump to
the write task.

It looks like the separation between the conn_write and
conn_mwrite task is slowly becoming obsolete and could
be replaced with a single task.

Change-Id: I5020a16bf8013a9349b8690d0d6b8e7a52789342
Reviewed-on: http://review.couchbase.org/79100
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Will Gardner <willg@rdner.io>