MB-20940: Remove ship_mcbp_tap_log
TAP is going to to be removed (and is currently disabled by not
allowing anyone to connect to it). This patch nukes the
ship_mcbp_tap_log method as I don't want to spend time fixing
the buffer patch to work on top of TAP when we're going to kill it
anyway.

Change-Id: I66ce71115d9f4cc9ee7f9ef86949ea92010ac858
Reviewed-on: http://review.couchbase.org/79105
Reviewed-by: Will Gardner <willg@rdner.io>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>