MB-24474: RangeIteratorLL to reset 'readRange' only when it holds 'readLock'
When a RangeIteratorLL is done reading all the items, it resets the
readRange and releases the linked list rangeReadLock. The readRange
is also reset in the destructor, and this is erroneous as the
rangeReadLock might have been released before this (as mentioned
above).

This commit fixes the problem by resetting the readRange in the
destructor only if the rangeReadLock is held.

Also a test case is added for the same.

Change-Id: Ie3bb55cff2bbfd65f08677cf823aa5d60d37a779
Reviewed-on: http://review.couchbase.org/79073
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>