MB-24432: Store a seqno to _local to determine when a CAS is a HLC value
Add a new field to _local/vbstate called hlc_epoch, this is a seqno
which we will later use to determine if a value has a CAS that is
a HLC (as opposed to a CAS generated in 3.x).

This persisted seqno will be used in the future to determine which
items can be tagged with a last-update virtual xattr (i.e. items we
know that the CAS is HLC generated as opposed to a CAS generated by
3.x).

Change-Id: I5ef5a8a08b06e69e3160ed71524f74c3053fd59e
Reviewed-on: http://review.couchbase.org/79127
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>