MB-24422: Avoid race condition in access scanner start
The Access Scanner config listeners are added only after the warmup is
complete. The atomic variable tracking the warmup state is first set to
be DONE and only then the listeners are added.There is small time gap
between these two events.But the tests wait for warmup to be completed
based on the atomic var. If the config is changed during this gap, that
the change is not reflected on to the scanner. So we have to set it
again for the change listeners to trigger and then set the
stats.AlogTime value from the config.alog_task_time correctly.

Change-Id: I453fc4ca2c4e60daab5bcdac367adbc50d2345e2
Reviewed-on: http://review.couchbase.org/79162
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>